pyproject.toml
[tool.poetry]
name = "web-scrapping-service"
version = "1.0.0"
description = "A simple web scraping API service inspired by ScrapingBee"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "*.py"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
playwright = "^1.40.0"
beautifulsoup4 = "^4.12.2"
lxml = "^4.9.3"
redis = "^5.0.1"
python-dotenv = "^1.0.0"
pydantic = "^2.5.0"
httpx = "^0.25.2"
aiohttp = "^3.9.1"
slowapi = "^0.1.9"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "main:main"


